split skate.d.ts into schedule.d.ts and realtime.d.ts
