+++ assets/src/components/incomingBox.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from "react"
 import DispatchContext from "../contexts/dispatchContext"
 import { directionOnLadder, VehicleDirection } from "../models/ladderVehicle"
 import { status } from "../models/vehicleStatus"
-import { Vehicle, VehicleId } from "../skate"
+import { Vehicle, VehicleId } from "../realtime.d"
 import { selectVehicle } from "../state"
 import { LadderDirection } from "./ladder"
 import VehicleIcon, { Orientation, Size } from "./vehicleIcon"
