+++ assets/src/components/vehiclePropertiesPanel.tsx
@@ -2,7 +2,8 @@ import React, { useContext, useEffect } from "react"
 import DispatchContext from "../contexts/dispatchContext"
 import detectSwipe, { SwipeDirection } from "../helpers/detectSwipe"
 import { status } from "../models/vehicleStatus"
-import { DataDiscrepancy, Route, Vehicle } from "../skate.d"
+import { DataDiscrepancy, Vehicle } from "../realtime.d"
+import { Route } from "../schedule.d"
 import { deselectVehicle } from "../state"
 import CloseButton from "./closeButton"
 import Map from "./map"
