+++ assets/src/hooks/useRoutes.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react"
 import { fetchRoutes } from "../api"
-import { Route } from "../skate"
+import { Route } from "../schedule.d"
 
 const useRoutes = (): Route[] | null => {
   const [routes, setRoutes] = useState<Route[] | null>(null)
